+++ test/e2e/suite/mergerequests_meta_test.go
@@ -72,12 +72,12 @@ func TestMeta_MRDeep(t *testing.T) {
 		requireNoError(t, err, "create group for list_group")
 		defer func() {
 			_ = callToolVoidOn(ctx, sess.meta, "gitlab_group", map[string]any{
-				"action": "delete", "params": map[string]any{"group_id": strconv.FormatInt(int64(grp.ID), 10)},
+				"action": "delete", "params": map[string]any{"group_id": strconv.FormatInt(grp.ID, 10)},
 			})
 		}()
 		out, err := callToolOn[mergerequests.ListOutput](ctx, sess.meta, "gitlab_merge_request", map[string]any{
 			"action": "list_group",
-			"params": map[string]any{"group_id": strconv.FormatInt(int64(grp.ID), 10)},
+			"params": map[string]any{"group_id": strconv.FormatInt(grp.ID, 10)},
 		})
 		requireNoError(t, err, "list_group")
 		t.Logf("Listed %d group MRs", len(out.MergeRequests))
